REFACTOR: move types into matching files
