+++ src/LiberisLabs.CompaniesHouse/Response/CompanyProfile/AccountingReferenceDateInfo.cs
@@ -1,0 +1,15 @@
+using Newtonsoft.Json;
+
+namespace LiberisLabs.CompaniesHouse.Response.CompanyProfile
+{
+    public class AccountingReferenceDateInfo
+    {
+        [JsonProperty(PropertyName ="day")]
+        public int Day { get; set; }
+
+        [JsonProperty(PropertyName = "month")]
+        public int Month { get; set; }
+
+        public string DayAndMonth => $"{Day}/{Month}";
+    }
+}
