+++ src/LiberisLabs.CompaniesHouse/Response/CompanyProfile/AccountsType.cs
@@ -1,9 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Runtime.Serialization;
-using System.Text;
-using System.Threading.Tasks;
+﻿using System.Runtime.Serialization;
 
 namespace LiberisLabs.CompaniesHouse.Response.CompanyProfile
 {
