+++ src/LiberisLabs.CompaniesHouse/Response/CompanyProfile/OfficerSummaryInfo.cs
@@ -1,0 +1,16 @@
+﻿using Newtonsoft.Json;
+
+namespace LiberisLabs.CompaniesHouse.Response.CompanyProfile
+{
+    public class OfficerSummaryInfo
+    {
+        [JsonProperty(PropertyName ="active_count")]
+        public int ActiveCount { get; set; }
+
+        [JsonProperty(PropertyName ="resigned_count")]
+        public int ResignedCount { get; set; }
+
+        [JsonProperty(PropertyName ="officers")]
+        public OfficerInfo[] Officers { get; set; }
+    }
+}
